\name{histogram}
\alias{histogram}

\title{
Ratings histogram.
}
\description{
Histogram of the ratings grouped by value.
}
\usage{
histogram(data, title = "", x = "Rating values", y = "# of ratings")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
class \code{"_ds"}, the dataset.
}
  \item{title}{
class \code{"character"}, eventual caption of for the chart.
}
  \item{x}{
class \code{"character"}, label for the x-axis.
}
  \item{y}{
class \code{"character"}, label for the y-axis.
}
}
